+++ changelog.md
@@ -3,10 +3,11 @@
 ## Upcoming
 ### Additions
 * Roofed Forest Vegetation data pack
-  * Lowers the giant mushrooms generated in a Roofed forests as well as removes the non-dark oak trees
+  * Lowers the giant mushrooms generated in Roofed forests as well as removes the non-dark oak trees
 
 ### Changes
-* Chorus Fruits are now fast eating and will cancel fall damage when teleporting
+* Cyan flower no longer generates naturally; instead, it's dug up by the Sniffer
+* horus Fruits are now fast eating and will cancel fall damage when teleporting
 
 ### Fixes
 * Advancements